@@ -1,10 +1,11 @@
-import React from "react";
+import { StrictMode } from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
 import App from "./App";
 import { store } from "./store";
 import { Provider } from "react-redux";
-import * as serviceWorker from "./serviceWorker";
+import * as serviceWorkerRegistration from "./serviceWorkerRegistration";
+import reportWebVitals from "./reportWebVitals";
 import {
   serviceWorkerInitialized,
   serviceWorkerUpdated,
@@ -12,22 +13,27 @@ import {
 import { HelmetProvider } from "react-helmet-async";
 
 ReactDOM.render(
-  <React.StrictMode>
+  <StrictMode>
     <Provider store={store}>
       <HelmetProvider>
         <App />
       </HelmetProvider>
     </Provider>
-  </React.StrictMode>,
+  </StrictMode>,
   document.getElementById("root")
 );
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.register({
+// Learn more about service workers: https://cra.link/PWA
+serviceWorkerRegistration.register({
   onSuccess: (registration) =>
     store.dispatch(serviceWorkerInitialized(registration)),
   onUpdate: (registration) =>
     store.dispatch(serviceWorkerUpdated(registration)),
 });
+
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+reportWebVitals();
