@@ -1,16 +1,13 @@
 import useAnswers from "hooks/useAnswers";
-import React, { useMemo } from "react";
+import { useMemo } from "react";
 import { Helmet } from "react-helmet-async";
 import { Link } from "react-router-dom";
 import { questionsPerPage } from "utils/constants";
 import styles from "./Homepage.module.css";
 
 const Homepage = () => {
-  const {
-    unansweredQuestions,
-    questionsWithAnswers,
-    clearAnswers,
-  } = useAnswers();
+  const { unansweredQuestions, questionsWithAnswers, clearAnswers } =
+    useAnswers();
   const [firstUnansweredQuestion] = unansweredQuestions;
   const hasStarted = useMemo(
     () =>
